Fix compile warnings on linux 64-bit

+++ tdishr/TdiCall.c
@@ -142,7 +142,7 @@ unsigned char			origin[255];
 	MdsFree1Dx(&entry, NULL);
 	MdsFree1Dx(&image, NULL);
 
-	newdsc[0] = (struct descriptor *)(narg-2);
+        newdsc[0] = (struct descriptor *)(long)(narg-2);
 	for (j = 2; j < narg && status & 1; ++j) {
 		for (pfun = (struct descriptor_function *)list[j];
                              pfun && pfun->dtype == DTYPE_DSC;)
